fixed and refactored digital services screen
