+++ mobile-app/app/src/main/java/com/semirsuljevic/raiffaisenmobileapp/ui/navigation/navigator.kt
@@ -19,7 +19,6 @@ import androidx.navigation.compose.composable
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.semirsuljevic.raiffaisenmobileapp.ui.composables.CenteredTitleAppBar
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.CurrentAccountScreen
-import com.semirsuljevic.raiffaisenmobileapp.ui.screens.DigitalScreensScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.Locations
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.SavingsScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_info.*
@@ -28,6 +27,7 @@ import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.LoginScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_more.*
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products.AccountSetsScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products.CreditCardsScreen
+import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products.DigitalScreensScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products.LoansScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.locations_filter.LocationsFilterScreen
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Gray200
