+++ mobile-app/app/src/main/java/com/semirsuljevic/raiffaisenmobileapp/ui/screens/intro_navbar/intro_products/DigitalServicesScreen.kt
@@ -1,16 +1,16 @@
-package com.semirsuljevic.raiffaisenmobileapp.ui.screens
+package com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
@@ -20,6 +20,7 @@ import com.semirsuljevic.raiffaisenmobileapp.R
 import com.semirsuljevic.raiffaisenmobileapp.ui.composables.BulletListItem
 import com.semirsuljevic.raiffaisenmobileapp.ui.composables.CenteredTitleAppBar
 import com.semirsuljevic.raiffaisenmobileapp.ui.composables.ListItemSeparator
+import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Black
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Gray200
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Yellow400
 
@@ -28,27 +29,33 @@ fun DigitalScreensScreen(navController: NavController) {
     val scrollState = rememberScrollState()
     val uriHandler = LocalUriHandler.current
     val moreInfoLink = stringResource(id = R.string.digital_services_more_info_link)
-    Column (Modifier.verticalScroll(scrollState)){
-        CenteredTitleAppBar(
-            title = stringResource(id = R.string.digital_services_screen_title),
-            navController = navController
-        )
-        ListItemSeparator()
-        Spacer(modifier = Modifier.height(20.dp))
-        Column(Modifier.padding(horizontal = 20.dp)) {
-            Text(
-                text = stringResource(id = R.string.digital_services_screen_description),
-                color = Gray200,
-                textAlign = TextAlign.Justify
+    Scaffold (
+        backgroundColor = Black,
+        topBar = {
+            CenteredTitleAppBar(
+                title = stringResource(id = R.string.digital_services_screen_title),
+                navController = navController
             )
+        }
+    ){
+        Column (Modifier.verticalScroll(scrollState)){
+            ListItemSeparator()
             Spacer(modifier = Modifier.height(20.dp))
-            DigitalServicesList()
-            Spacer(modifier = Modifier.height(40.dp))
-            Text(
-                text = stringResource(id = R.string.digital_services_more_info),
-                color = Yellow400,
-                modifier = Modifier.clickable { uriHandler.openUri(moreInfoLink) }
-            )
+            Column(Modifier.padding(horizontal = 20.dp)) {
+                Text(
+                    text = stringResource(id = R.string.digital_services_screen_description),
+                    color = Gray200,
+                    textAlign = TextAlign.Justify
+                )
+                Spacer(modifier = Modifier.height(10.dp))
+                DigitalServicesList()
+                Spacer(modifier = Modifier.height(40.dp))
+                Text(
+                    text = stringResource(id = R.string.digital_services_more_info),
+                    color = Yellow400,
+                    modifier = Modifier.clickable { uriHandler.openUri(moreInfoLink) }
+                )
+            }
         }
     }
 }
@@ -56,7 +63,9 @@ fun DigitalScreensScreen(navController: NavController) {
 
 @Composable
 fun DigitalServicesList() {
-    Column {
+    Column (
+        modifier = Modifier.padding(horizontal = 10.dp)
+    ){
         BulletListItem(text = stringResource(id = R.string.digital_services_item1))
         BulletListItem(text = stringResource(id = R.string.digital_services_item2))
         BulletListItem(text = stringResource(id = R.string.digital_services_item3))
